Crate actix_tracing

Actix tracing - support for tokio tracing with Actix services.

Structs

TracingService

A Service implementation that automatically enters/exits tracing spans for the wrapped inner service.

TracingTransform

A Transform implementation that wraps services with a TracingService.

Functions

trace

Wraps the provided service factory with a transform that automatically enters/exits the given span.